develop salary calculator form
