+++ src/components/salary_calculator/result_block.tsx
@@ -2,10 +2,12 @@ import React from 'react';
 import { numberWithCommas } from '@/lib/utils/common';
 import { RowItem } from '@/interfaces/calculator';
 
-const ResultBlock: React.FC<{
+interface IResultBlockProps {
   backgroundColor: string;
   rowItems: RowItem[];
-}> = ({ backgroundColor, rowItems }) => {
+}
+
+const ResultBlock: React.FC<IResultBlockProps> = ({ backgroundColor, rowItems }) => {
   // Info: (20250708 - Julian) 項目總計：取出 rowItems 的最後一個項目
   const totalItem = rowItems.slice(-1)[0];
   const displayTotalRowItem = totalItem && (
